+++ pkg/shell/encoding_unix.go
@@ -1,0 +1,10 @@
+//go:build !windows
+// +build !windows
+
+package shell
+
+import "golang.org/x/text/encoding"
+
+func windowsEncoding() encoding.Encoding {
+	return encoding.Nop
+}