+++ pkg/shell/host_test.go
@@ -1,8 +1,10 @@
-// nolint: errcheck
+// nolint: errcheck, dupl
 package shell_test
 
 import (
+	"bytes"
 	"context"
+	"errors"
 	"fisherman/pkg/guards"
 	. "fisherman/pkg/shell"
 	"fisherman/testing/mocks"
@@ -12,6 +14,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"golang.org/x/text/encoding"
+	"golang.org/x/text/transform"
 )
 
 func TestHost_Start(t *testing.T) {
@@ -40,7 +44,8 @@ func TestHost_Start(t *testing.T) {
 		command := exec.Command("echo", "test")
 		command.Stdin = strings.NewReader("test")
 		strategy := mocks.NewShellStrategyMock(t).
-			GetCommandMock.Return(command)
+			GetCommandMock.Return(command).
+			GetEncodingMock.Return(encoding.Nop)
 		host := NewHost(context.TODO(), strategy)
 		defer host.Terminate()
 
@@ -86,13 +91,65 @@ func TestHost_Write(t *testing.T) {
 		command := exec.Command("echo", "test")
 		command.Stdin = strings.NewReader("test")
 		strategy := mocks.NewShellStrategyMock(t).
-			GetCommandMock.Return(command)
+			GetCommandMock.Return(command).
+			GetEncodingMock.Return(encoding.Nop)
 		host := NewHost(context.TODO(), strategy)
 		defer host.Terminate()
 
 		_, err := fmt.Fprintln(host, "echo 'test'")
 
 		assert.EqualError(t, err, "exec: Stdin already set")
+	})
+
+	t.Run("write in call endoder", func(t *testing.T) {
+		encoderTransformer := mocks.NewTransformerMock(t).
+			ResetMock.Return().
+			TransformMock.Set(transform.Nop.Transform)
+
+		decoderTransformer := mocks.NewTransformerMock(t).
+			ResetMock.Return().
+			TransformMock.Set(transform.Nop.Transform)
+
+		encoding := mocks.NewEncodingMock(t).
+			NewDecoderMock.Return(&encoding.Decoder{Transformer: decoderTransformer}).
+			NewEncoderMock.Return(&encoding.Encoder{Transformer: encoderTransformer})
+
+		host := NewHost(context.TODO(), Default(), WithEncoding(encoding))
+		defer host.Terminate()
+
+		_, err := fmt.Fprintln(host, "echo 'test'")
+
+		_ = host.Wait()
+
+		assert.NoError(t, err)
+		assert.NotEmpty(t, encoderTransformer.TransformMock.Calls())
+		assert.Empty(t, decoderTransformer.TransformMock.Calls())
+	})
+
+	t.Run("write in call endoder and decoder", func(t *testing.T) {
+		encoderTransformer := mocks.NewTransformerMock(t).
+			ResetMock.Set(transform.Nop.Reset).
+			TransformMock.Set(transform.Nop.Transform)
+
+		decoderTransformer := mocks.NewTransformerMock(t).
+			ResetMock.Set(transform.Nop.Reset).
+			TransformMock.Set(transform.Nop.Transform)
+
+		encoding := mocks.NewEncodingMock(t).
+			NewDecoderMock.Return(&encoding.Decoder{Transformer: decoderTransformer}).
+			NewEncoderMock.Return(&encoding.Encoder{Transformer: encoderTransformer})
+
+		buff := &bytes.Buffer{}
+		host := NewHost(context.TODO(), Default(), WithEncoding(encoding), WithStdout(buff), WithStderr(buff))
+		defer host.Terminate()
+
+		_, err := fmt.Fprintln(host, "echo 'test'")
+
+		_ = host.Wait()
+
+		assert.NoError(t, err)
+		assert.NotEmpty(t, encoderTransformer.TransformMock.Calls())
+		assert.NotEmpty(t, decoderTransformer.TransformMock.Calls())
 	})
 }
 
@@ -130,6 +187,28 @@ func TestHost_Close(t *testing.T) {
 			assert.NoError(t, err)
 		})
 	})
+
+	t.Run("encoding error", func(t *testing.T) {
+		transformer := mocks.NewTransformerMock(t).
+			ResetMock.Set(transform.Nop.Reset).
+			TransformMock.Set(func(dst, src []byte, atEOF bool) (nDst int, nSrc int, err error) {
+			return 0, 0, errors.New("encoding error")
+		})
+
+		encoding := mocks.NewEncodingMock(t).
+			NewDecoderMock.Return(&encoding.Decoder{Transformer: transformer}).
+			NewEncoderMock.Return(&encoding.Encoder{Transformer: transformer})
+
+		host := NewHost(context.TODO(), Default(), WithEncoding(encoding))
+		defer host.Terminate()
+
+		_ = host.Start()
+
+		actual := host.Close()
+
+		assert.EqualError(t, actual, "encoding error")
+		assert.NotEmpty(t, transformer.TransformMock.Calls())
+	})
 }
 
 func TestHost_Terminate(t *testing.T) {
@@ -148,5 +227,28 @@ func TestHost_Terminate(t *testing.T) {
 		err := host.Terminate()
 
 		assert.NoError(t, err)
+	})
+}
+
+func TestHost_Wait(t *testing.T) {
+	t.Run("return encoding error", func(t *testing.T) {
+		transformer := mocks.NewTransformerMock(t).
+			ResetMock.Set(transform.Nop.Reset).
+			TransformMock.Set(func(dst, src []byte, atEOF bool) (nDst int, nSrc int, err error) {
+			return 0, 0, errors.New("encoding error")
+		})
+
+		encoding := mocks.NewEncodingMock(t).
+			NewDecoderMock.Return(&encoding.Decoder{Transformer: transformer}).
+			NewEncoderMock.Return(&encoding.Encoder{Transformer: transformer})
+
+		host := NewHost(context.TODO(), Default(), WithEncoding(encoding))
+		defer host.Terminate()
+		_ = host.Start()
+
+		actual := host.Wait()
+
+		assert.EqualError(t, actual, "encoding error")
+		assert.NotEmpty(t, transformer.TransformMock.Calls())
 	})
 }